Add ability for flags/switches/arguments to "network" (#40)
This PR will introduce arguments for the `network` command:
- `-a`, `--active` to show active networks
- `-n`, `--name` to show network names
- `-d`, `--device` to show network devices
- `-v6`, `--ipv6` to show ipv6 addresses
- `-v8`, `--ipv8` to show ipv8 addresses

It also safeguards against invalid arguments, and lets the user know if they provide one, and gives them a clue to use the `help network` command to get more information about flags.

This PR will also fix executable file saving and parsing, so the `help` command shows all the data properly. The issue was that there was a split on `\n` for lines, but values in the help text could contain `\n`. Similarly, separating keys and values for help was using `:`, but `:` could be included in the help values. Those separator characters have been updated to more specific strings that will not show up in help information.